[🐛 fix] 카카오 소셜 로그인 수정

+++ src/main/java/speedit/bookplate/oAuth/dto/PostOauthRes.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @Builder
 @AllArgsConstructor
 public class PostOauthRes {
-    private int userIdx;
+    private long userIdx;
     private String jwt;
 }
